registrationtools/requirements.txt
PythonHelper.egg==info
numpy==1.13.1
SimpleITK==1.0.1
